+++ include/aww-common.hpp
@@ -14,6 +14,7 @@
 #include <chrono>
 #include <thread>
 #include <stdexcept>
+#include <string_view>
 
 // Windows tricks for aww::os::Proccess
 #ifdef _WIN32
@@ -24,6 +25,25 @@
 
 namespace aww
 {
+  // CallTag struct definition
+  struct call_tag_t {
+    constexpr explicit call_tag_t(unsigned long long value) : value(value) {}
+
+    const unsigned long long value;
+  };
+
+  // Compile-time hash function
+  constexpr unsigned long long _compiletime_hash(const char* str, unsigned long long hash = 0, size_t index = 0) {
+    return str[index] ? _compiletime_hash(str, (hash * 131) + str[index], index + 1) : hash;
+  }
+
+  // call_tag function with compile-time length check for string literals
+  template <size_t N>
+  constexpr call_tag_t call_tag(const char (&str)[N]) {
+      static_assert(N > 11, "Tag string must be at least 11 characters long."); // N includes the null terminator
+      return call_tag_t(_compiletime_hash(str));
+  }
+
   class Result
   {
   public: